% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{correlation_analysis}
\alias{correlation_analysis}
\title{Generating a correlation matrix of pathways and TFs}
\usage{
correlation_analysis(
  tf_data,
  pathway_data,
  render_plot = TRUE,
  return_corr_data = FALSE
)
}
\arguments{
\item{tf_data}{transcription factor data by cell}

\item{pathway_data}{pathway data by cell}

\item{render_plot}{render plot? default = TRUE}

\item{return_corr_data}{return the correlation data? default = FALSE}
}
\description{
Takes cell-wise data from the 'scores_handling' functions and generates a correlation matrix
}
\examples{
correlation_analysis(tf_data_bycell, pathway_data_bycell, return_corr_data=FALSE, render_plot=TRUE)
}
\keyword{correlation}
\keyword{plot}
