% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_calculations.R
\name{custom_pathway_calc}
\alias{custom_pathway_calc}
\title{Calculate pathway activity using a dataframe of custom pathways}
\usage{
custom_pathway_calc(
  seurat_obj,
  pathways,
  num_genes,
  organism = "Human",
  assay_name
)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{pathways}{dataframe of custom pathways (must be same format as PROGENy pathways)}

\item{num_genes}{number of top genes to return for each pathway if full_model = FALSE.}

\item{organism}{"Human" or "Mouse", defaults to "Human".}

\item{assay_name}{string to name the resulting Seurat object assay}
}
\description{
This function uses the PROGENy framework to quantify pathway activity given a dataframe of custom pathways and returns a seurat object assay with the results.
}
\examples{
custom_pathway_calc(seurat_obj, pathways, 500, "Human", "custom_pathway_activity"))
}
\keyword{activity}
\keyword{calculation}
\keyword{custom}
\keyword{pathway}
