% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dorothea.R
\name{custom_regulons_calc}
\alias{custom_regulons_calc}
\title{Quantifying TF activity using a data frame of custom regulons and the VIPER algorithm}
\usage{
custom_regulons_calc(seurat_obj, regulons, assay_name, cores = 16)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{regulons}{data frame of custom regulons (must be same format as DoRothEA regulons)}

\item{assay_name}{string to name the resulting Seurat object assay}

\item{cores}{number of cores to use for calculation (default is 16)}
}
\description{
This function uses VIPER to quantify TF activity given a data frame of custom regulons and returns a Seurat object assay with the results.
}
\examples{
custom_regulons_calc(data,regulons,"custom_TFactivity"))
}
\keyword{Quantifying}
\keyword{TF}
\keyword{VIPER}
\keyword{activity}
\keyword{algorithm}
\keyword{and}
\keyword{custom}
\keyword{regulons}
\keyword{the}
\keyword{with}
