% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorothea_scores_handling.R
\name{dorothea_scores_handling}
\alias{dorothea_scores_handling}
\title{Handling the dorothea scores (proportion adjustments, summarizing by cell type, etc.)}
\usage{
dorothea_scores_handling(seurat_obj, comparison_feature, topTFs)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{comparison_feature}{which metadata do you want to do comparisons for?}

\item{topTFs}{number of top TFs to be us ed for heatmap downstream}
}
\description{
This function takes a seurat object and produces a list of dataframes that are necessary for downstream analysis
}
\examples{
dorothea_scores_handling(seurat_obj,seurat_obj$meta.data$cell_type, 30)
}
\keyword{dorothea}
\keyword{handling}
\keyword{score}
