% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorothea_calculations.R
\name{dorothea_tf_calc}
\alias{dorothea_tf_calc}
\title{Calculating VIPER scores (TF Activity) from DoRothEA Regulons}
\usage{
dorothea_tf_calc(seurat_obj, conf_scores)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{conf_scores}{vector of DoRothEA confidence scores (A,B,C,D,E)}
}
\description{
This function takes a seurat object and a vector of DoRothEA confidence levels and returns the same Seurat object with a "dorothea" assay.
}
\examples{
dorothea_tf_calc(seurat_obj, c("A","B","C","D"))
}
\keyword{VIPER}
\keyword{calculations}
