% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_heatmap.R
\name{downstream_heatmap}
\alias{downstream_heatmap}
\title{Plotting dorothea/progeny scores by comparison group on a heatmap}
\usage{
downstream_heatmap(df, title)
}
\arguments{
\item{title}{title for the heatmap}

\item{data}{Data comes from the scores_handling functions, plot the top TFs or pathways}
}
\description{
This function takes data from the 'scores_handling' functions and plots a heatmap of the top TFs or pathways
}
\examples{
downstream_heatmap(data, "top 30 TFs, by indication")
}
\keyword{downstream}
\keyword{heatmap}
