% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_associated_pathways.R
\name{find_associated_pathways}
\alias{find_associated_pathways}
\title{Given a list of genes, find which pathways they are associated with and see how statistically significant those pathways are for a given comparison (e.g. healthy vs disease)}
\usage{
find_associated_pathways(
  gene_vector,
  pathway_data_bygroup = NULL,
  effect_size_data = NULL,
  custom_pathways = NULL,
  progeny_pathways = TRUE,
  num_genes = 500,
  organism = "Human"
)
}
\arguments{
\item{gene_vector}{vector of genes}

\item{pathway_data_bygroup}{pathway data summarized by comparison group}

\item{effect_size_data}{dataframe of cohen's D scores for a given comparison (optional, only if comparison is 2 groups)}

\item{custom_pathways}{searches custom pathways for genes in gene vector, default is FALSE.}
}
\description{
This function takes a vector of genes and pathway activity data (by group, not by cell...optional) and a data frame of cohen's D scores (if the comparison is 2 groups...optional).
If only the gene vector is supplied, the function will return all pathways for which a given gene is part of the pathway's gene set.
}
\examples{
find_associated_pathways(gene_vector, pathway_data_bygroup, effect_size_data, custom_pathways=custom_pathways, progeny_pathways=FALSE)
}
\keyword{TFs}
\keyword{associated}
\keyword{find}
