% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_associated_TFs.R
\name{find_associated_TFs}
\alias{find_associated_TFs}
\title{Given a list of genes, find which TFs they are associated with and see how statistically significant those TFs are for a given comparison (e.g. healthy vs disease)}
\usage{
find_associated_TFs(
  gene_vector,
  tf_data_bygroup = NULL,
  effect_size_data = NULL,
  custom_regulons = NULL
)
}
\arguments{
\item{gene_vector}{vector of genes}

\item{tf_data_bygroup}{TF data summarized by comparison group}

\item{effect_size_data}{dataframe of cohen's D scores for a given comparison (optional, only if comparison is 2 groups)}

\item{custom_regulons}{searches custom TF regulons for genes in gene vector, default is FALSE.}
}
\description{
This function takes a vector of genes and TF activity data (by group, not by cell...optional) and a dataframe of cohen's D scores (if the comparison is 2 groups...optional).
If only the gene vector is supplied, the function will return all TFs for which a given gene is part of the regulon for the entire dorothea collection of TFs.
}
\examples{
find_associated_TFs(gene_vector, tf_data_bygroup, effect_size_data, custom_regulons=custom_regulons)
}
\keyword{TFs}
\keyword{associated}
\keyword{find}
