% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dorothea.R
\name{get_dorothea_regulons}
\alias{get_dorothea_regulons}
\title{Returning DoRothEA Regulons}
\usage{
get_dorothea_regulons(conf_scores)
}
\arguments{
\item{conf_scores}{vector of DoRothEA confidence scores (A,B,C,D,E)}
}
\description{
Returns DoRothEA regulons from a vector of DoRothEA confidence levels.
}
\examples{
get_dorothea_regulons(c("A","B","C","D"))
}
\keyword{DoRothEA}
\keyword{Regulon}
