% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_calculations.R
\name{get_progeny_pathways}
\alias{get_progeny_pathways}
\title{Returning PROGENy pathways}
\usage{
get_progeny_pathways(full_model = TRUE, num_genes = NULL, organism = "Human")
}
\arguments{
\item{full_model}{return the full model? Defaults to TRUE.}

\item{num_genes}{number of top genes to return for each pathway if full_model = FALSE.}

\item{organism}{"Human" or "Mouse", defaults to "Human".}
}
\description{
This function returns PROGENy pathways
}
\examples{
get_progeny_pathways(full_model=FALSE, num_genes=100, organism = "Human")
}
\keyword{PROGENy}
\keyword{pathways}
