% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_scores_handling.R
\name{handle_progeny_scores}
\alias{handle_progeny_scores}
\title{Handling the progeny scores (proportion adjustments, summarizing by cell type, etc.)}
\usage{
handle_progeny_scores(seurat_obj, comparison_feature)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{comparison_feature}{which metadata feature do you want to do comparisons for?}
}
\description{
Takes a Seurat object and produces a list of data frames that are necessary for downstream analysis
}
\examples{
handle_progeny_scores(pnmc10k,30)
}
\keyword{handling}
\keyword{progeny}
\keyword{score}
