% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_effectsize_calculations.R
\name{pathway_effsize_calc}
\alias{pathway_effsize_calc}
\title{Cohen's D effect size calculation for progeny data (requires 2 group comparisons...e.g. healthy vs disease)}
\usage{
pathway_effsize_calc(pathway_data_bycell)
}
\arguments{
\item{pathway_data_bycell}{dataframe of pathway data by cell}
}
\description{
This function takes progeny (or more generally pathway activity) data by cell and produces a dataframe of cohen's D scores for each pathway.
}
\examples{
pathway_effsize_calc(pathway_data_bycell)
}
\keyword{calculation}
\keyword{effect}
\keyword{progeny}
\keyword{size}
