% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooledSD_calculation.R
\name{pooledSD}
\alias{pooledSD}
\title{Pooled Standard Deviation Calculation}
\usage{
pooledSD(sdvec)
}
\arguments{
\item{sdvec}{vector of standard deviation values.}
}
\description{
This function allows you to calculate a pooled standard deviation from a vector of standard deviations.
}
\examples{
pooledSD()
}
\keyword{pooledSD}
