% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_calculations.R
\name{progeny_score_calc}
\alias{progeny_score_calc}
\title{Calculating pathway activity scores from PROGENy pathways}
\usage{
progeny_score_calc(seurat_obj, num_genes, organism = "Human")
}
\arguments{
\item{seurat_obj}{seurat object}

\item{num_genes}{number of pathway genes to use for calculations}

\item{organism}{"Human" or "Mouse", defaults to "Human"}
}
\description{
This function takes a seurat object and a number of pathway genes and returns the same Seurat object with a "progeny" assay.
}
\examples{
progeny_score_calc(seurat_obj, 500, "Human")
}
\keyword{PROGENy}
\keyword{calculations}
