% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_scores_handling.R
\name{progeny_scores_handling}
\alias{progeny_scores_handling}
\title{Handling the progeny scores (proportion adjustments, summarizing by cell type, etc.)}
\usage{
progeny_scores_handling(seurat_obj, comparison_feature)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{comparison_feature}{which metadata do you want to do comparisons for?}
}
\description{
This function takes a seurat object and produces a list of dataframes that are necessary for downstream analysis
}
\examples{
progeny_scores_handling(seurat_obj,30)
}
\keyword{handling}
\keyword{progeny}
\keyword{score}
