% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_progeny_pathways.R
\name{reformat_progeny_pathways}
\alias{reformat_progeny_pathways}
\title{Reformat the progeny pathway data when you select only the top n genes}
\usage{
reformat_progeny_pathways(pathway_data)
}
\arguments{
\item{pathway_data}{progeny pathway data for top n genes (from \code{get_progeny_pathways()} function)}
}
\description{
When selecting progeny pathway data and choosing only a subset of the top genes from each pathway, the output is a sparse matrix and is not conducive for downstream analysis. This reformats to a long data frame.
}
\examples{
reformat_progeny_pathways(pathway_data = top500genes_progenypathways)
}
\keyword{pathways}
\keyword{progeny}
\keyword{reformat}
