% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pySCENIC.R
\name{run_cistarget}
\alias{run_cistarget}
\title{Uses reticulate package to run CisTarget (regulon prediction) from R}
\usage{
run_cistarget(dir, loom_path, rank_db_path, motif_path)
}
\arguments{
\item{dir}{directory path to output pySCENIC results}

\item{loom_path}{path to write initial loom file}

\item{rank_db_path}{path to ranking database}

\item{motif_path}{path to motif list}
}
\description{
Run pySCENIC pipeline
}
\examples{
run_cistarget(dir, loom_path, rank_db_path, motif_path)
}
\keyword{CisTarget}
