% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dorothea.R
\name{run_dorothea}
\alias{run_dorothea}
\title{Quantifying TF activity from DoRothEA Regulons}
\usage{
run_dorothea(seurat_obj, conf_scores, cores = 16)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{conf_scores}{vector of DoRothEA confidence scores (A,B,C,D,E)}

\item{cores}{number of cores to use for calculation (default = 16)}
}
\description{
Takes a Seurat object and a vector of DoRothEA confidence levels and returns the same Seurat object with a "dorothea" assay of TF activity scores.
}
\examples{
run_dorothea(pbmc10k, c("A","B","C","D"),cores=20)
}
\keyword{Activity}
\keyword{DoRothEA}
\keyword{Regulons}
\keyword{TF}
\keyword{with}
