% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pySCENIC.R
\name{run_grn}
\alias{run_grn}
\title{Uses reticulate package to run GRN from R}
\usage{
run_grn(dir, loom_path, tfs_path)
}
\arguments{
\item{dir}{directory path to output pySCENIC results}

\item{loom_path}{path to write initial loom file}

\item{tfs_path}{path to list of TFs}
}
\description{
Run GRN from pySCENIC pipeline
}
\examples{
run_grn(dir, loom_path, tfs_path,)
}
\keyword{GRN}
