% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pySCENIC.R
\name{run_loom_setup}
\alias{run_loom_setup}
\title{Uses reticulate package to set up loom file for pySCENIC pipeline}
\usage{
run_loom_setup(anndata_path, loom_path)
}
\arguments{
\item{anndata_path}{path to anndata object}

\item{loom_path}{path to write initial loom file}
}
\description{
Run GRN from pySCENIC
}
\examples{
run_loom_setup(anndata_path, loom_path)
}
\keyword{loom}
\keyword{pySCENIC}
\keyword{set}
\keyword{up}
