% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pySCENIC.R
\name{run_pyscenic}
\alias{run_pyscenic}
\title{Uses reticulate package to run pySCENIC from R}
\usage{
run_pyscenic(
  dir,
  anndata_path,
  loom_path,
  tfs_path,
  rank_db_path,
  motif_path,
  output_loom_path
)
}
\arguments{
\item{dir}{directory path to output pySCENIC results}

\item{anndata_path}{path to anndata object}

\item{loom_path}{path to write initial loom file}

\item{tfs_path}{path to list of TFs}

\item{rank_db_path}{path to ranking database}

\item{motif_path}{path to motif list}

\item{output_loom_path}{path to write output loom file}
}
\description{
Run pySCENIC pipeline
}
\examples{
run_pyscenic(outdir, anndata_path, loom_path, tfs_path, rank_db_path, motif_path, output_loom_path)
}
\keyword{pySCENIC}
