% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRegNetwrapper.R
\name{scRegNetwrapper}
\alias{scRegNetwrapper}
\title{Wrapper function to run dorothea and progeny and get basic downstream results}
\usage{
scRegNetwrapper(
  seurat_obj,
  dorothea_conf_scores = c("A", "B", "C", "D"),
  topTFs = 30,
  cores = 16,
  top_progeny_genes = 500,
  organism = "Human",
  comparison_feature
)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{dorothea_conf_scores}{vector of DoRothEA confidence scores (default = A->D)}

\item{topTFs}{topTFs for downstream heatmap (default = 30)}

\item{cores}{number of cores to use for calculation (default = 16)}

\item{top_progeny_genes}{number of genes per progeny pathway to use for calculations (default = 500)}

\item{organism}{for progeny, use Human or Mouse pathways (default = "Human")}

\item{comparison_feature}{which metadata do you want to do comparisons for?}
}
\description{
Takes a seurat object an
}
\examples{
scRegNetwrapper(seurat_obj = pbmc10k, dorothea_conf_scores = c("A","B","C","D"), topTFs = 30, cores = 20, top_progeny_genes = 100, organism = "Human", comparison_feature = Idents(pbmc10k))
}
\keyword{all-in-one}
\keyword{and}
\keyword{dorothea}
\keyword{for}
\keyword{function}
\keyword{progeny}
\keyword{wrapper}
