% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorothea_effectsize_calculations.R
\name{tf_effsize_calc}
\alias{tf_effsize_calc}
\title{Cohen's D effect size calculation for dorothea data (requires 2 group comparisons...e.g. healthy vs disease)}
\usage{
tf_effsize_calc(tf_data_bycell)
}
\arguments{
\item{tf_data_bycell}{dataframe of TF data by cell}
}
\description{
This function takes a dorothea (or more generally tf activity) data by cell and produces a dataframe of cohen's D scores for each TF.
}
\examples{
tf_effsize_calc(tf_data_bycell=tf_data_bycell)
}
\keyword{calculation}
\keyword{dorothea}
\keyword{effect}
\keyword{size}
