% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorothea_calculations.R
\name{custom_regulons_calc}
\alias{custom_regulons_calc}
\title{Using VIPER to calculate TF activity using a data frame of custom regulons}
\usage{
custom_regulons_calc(seurat_obj, regulons, assay_name)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{regulons}{dataframe of custom regulons (must be same format as DoRothEA regulons)}

\item{assay_name}{string to name the resulting Seurat object assay}
}
\description{
This function uses VIPER to quantify TF activity given a data frame of custom regulons and returns a Seurat object assay with the results.
}
\examples{
custom_regulons_calc(data,regulons,"custom_TFactivity"))
}
\keyword{TF}
\keyword{custom}
\keyword{regulons}
