% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_progeny.R
\name{get_progeny_pathways}
\alias{get_progeny_pathways}
\title{Returning PROGENy pathways}
\usage{
get_progeny_pathways(num_genes = NULL, organism = "Human")
}
\arguments{
\item{num_genes}{number of top genes to return for each pathway if full_model = FALSE.}

\item{organism}{"Human" or "Mouse", defaults to "Human".}

\item{full_model}{return the full model? Defaults to TRUE.}
}
\description{
This function returns PROGENy pathways
}
\examples{
get_progeny_pathways(num_genes=100, "Human")
}
\keyword{PROGENy}
\keyword{pathways}
