% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pySCENIC.R
\name{run_cistarget}
\alias{run_cistarget}
\title{Uses reticulate package to run CisTarget (regulon prediction) from R}
\usage{
run_cistarget(
  dir,
  adjacencies_fname,
  loom_path,
  rank_db_path,
  motif_path,
  regulons_fname
)
}
\arguments{
\item{dir}{directory path to output pySCENIC results}

\item{adjacencies_fname}{filename of adjacencies output}

\item{loom_path}{path to initial loom file}

\item{rank_db_path}{path to ranking database}

\item{motif_path}{path to motif list}

\item{regulons_fname}{filename for outputting regulons}
}
\description{
Run pySCENIC pipeline
}
\examples{
run_cistarget(dir, adjacencies_fname, loom_path, rank_db_path, motif_path, regulons_fname)
}
\keyword{CisTarget}
